+++ packages/core/lib/inspect.js
@@ -1,0 +1,1 @@
+module.exports = require('q-i').print