+++ packages/core/CHANGELOG.md
@@ -1,6 +1,6 @@
 # @webpack-blocks/core - Changelog
 
-## Next version
+## 1.0.0
 
 - Added `when()` ([#242](https://github.com/andywer/webpack-blocks/issues/242))
 - Support negations in `match()` to set `exclude` ([#252](https://github.com/andywer/webpack-blocks/issues/252))